Removal and organization of data explorer code
- Moved most of the computes/methods to separate files
- Added comments for most of the functions
